Add a fast path for the tessellation of stroked curves.
