+++ crates/tessellation/src/geometry_builder.rs
@@ -474,12 +474,12 @@ where
         if a == b || a == c || b == c {
             println!("bad triangle {:?} {:?} {:?}", a, b, c);
         }
-        //debug_assert!(a != b);
-        //debug_assert!(a != c);
-        //debug_assert!(b != c);
-        //debug_assert!(a != VertexId::INVALID);
-        //debug_assert!(b != VertexId::INVALID);
-        //debug_assert!(c != VertexId::INVALID);
+        debug_assert!(a != b);
+        debug_assert!(a != c);
+        debug_assert!(b != c);
+        debug_assert!(a != VertexId::INVALID);
+        debug_assert!(b != VertexId::INVALID);
+        debug_assert!(c != VertexId::INVALID);
         self.buffers.indices.push((a + self.vertex_offset).into());
         self.buffers.indices.push((b + self.vertex_offset).into());
         self.buffers.indices.push((c + self.vertex_offset).into());
