+++ crates/tessellation/src/lib.rs
@@ -195,7 +195,6 @@ pub mod geometry_builder;
 mod math_utils;
 mod monotone;
 mod stroke;
-pub mod variable_stroke;
 mod basic_shapes;
 
 #[cfg(test)]
@@ -316,6 +315,22 @@ pub enum VertexSource {
         to: EndpointId,
         t: f32,
     },
+}
+
+impl VertexSource {
+    pub fn is_endpoint(&self) -> bool {
+        match self {
+            VertexSource::Endpoint { .. } => true,
+            _ => false,
+        }
+    }
+
+    pub fn is_edge(&self) -> bool {
+        match self {
+            VertexSource::Edge { .. } => true,
+            _ => false,
+        }
+    }
 }
 
 /// Vertical or Horizontal.
@@ -704,33 +719,4 @@ fn test_with_miter_limit() {
 #[should_panic]
 fn test_with_invalid_miter_limit() {
     let _ = StrokeOptions::default().with_miter_limit(0.0);
-}
-
-#[test]
-fn test_line_width() {
-    use crate::math::{point, Point};
-    let mut builder = crate::path::Path::builder();
-    builder.begin(point(0.0, 1.0));
-    builder.line_to(point(2.0, 1.0));
-    builder.end(false);
-    let path = builder.build();
-
-    let options = StrokeOptions::DEFAULT.with_line_width(2.0);
-    let mut geometry: VertexBuffers<Point, u16> = VertexBuffers::new();
-    StrokeTessellator::new()
-        .tessellate(
-            path.iter(),
-            &options,
-            &mut crate::geometry_builder::simple_builder(&mut geometry),
-        )
-        .unwrap();
-
-    for p in &geometry.vertices {
-        assert!(
-            *p == point(0.0, 0.0)
-                || *p == point(0.0, 2.0)
-                || *p == point(2.0, 0.0)
-                || *p == point(2.0, 2.0)
-        );
-    }
 }
